fix: move parseDur function to utils
