+++ src/commands/AboutCommand.ts
@@ -3,6 +3,7 @@ import Jukebox from "../structures/Jukebox";
 import { Message, MessageEmbed } from "discord.js";
 import { uptime as osUptime } from "os";
 import { version } from "discord.js";
+import { parseDur } from "../utils/parseDur";
 
 export default class PlayCommand extends BaseCommand {
     constructor(public client: Jukebox, readonly path: string) {
@@ -26,10 +27,10 @@ Playing Music on    :: ${this.client.guilds.cache.filter((g: any) => g.queue !==
 
 Platform            :: ${process.platform}
 Arch                :: ${process.arch}
-OS Uptime           :: ${this.parseDur(osUptime() * 1000)}
+OS Uptime           :: ${parseDur(osUptime() * 1000)}
 Memory              :: ${this.bytesToSize(Math.round(process.memoryUsage().rss))}
-Process Uptime      :: ${this.parseDur(Math.floor(process.uptime() * 1000))}
-Bot Uptime          :: ${this.parseDur(this.client.uptime!)}
+Process Uptime      :: ${parseDur(Math.floor(process.uptime() * 1000))}
+Bot Uptime          :: ${parseDur(this.client.uptime!)}
 
 NodeJS version      :: ${process.version}
 DiscordJS version   :: v${version}
@@ -38,25 +39,6 @@ Bot Version         :: v${require("../../package.json").version}
 Source code         :: https://sh.hzmi.xyz/jukebox
 \`\`\`
     `).setColor("#00FF00").setTimestamp());
-    }
-    private parseDur(ms: number): string {
-        let seconds = ms / 1000;
-        const days = parseInt((seconds / 86400).toString());
-        seconds = seconds % 86400;
-        const hours = parseInt((seconds / 3600).toString());
-        seconds = seconds % 3600;
-        const minutes = parseInt((seconds / 60).toString());
-        seconds = parseInt((seconds % 60).toString());
-
-
-        if (days) {
-            return `${days} days ${hours} hours ${minutes} minutes ${seconds} seconds`;
-        } else if (hours) {
-            return `${hours} hours ${minutes} minutes ${seconds} seconds`;
-        } else if (minutes) {
-            return `${minutes} minutes ${seconds} seconds`;
-        }
-        return `${seconds} seconds`;
     }
     private bytesToSize(bytes: number): string { // Function From Rendang's util (https://github.com/Hazmi35/rendang)
         if (isNaN(bytes) && bytes != 0) throw new Error(`[bytesToSize] (bytes) Error: bytes is not a Number/Integer, received: ${typeof bytes}`);
